kitchen: Initial guix dev environment
Modify the dependencies in the Gemfile to match what is in guix and
remove the Gemfile.lock. The lockfile is not required for early dev
and should be locked by the guix manifest instead.

Also add a guix manifest that can be used for development with:

guix shell --manifest=./config.scm

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>