kitchen: Migrate to gem
Move the application code into the 'lib' directory and fix up all the
paths internal to the codebase. Add a gemspec controlling the
installation from the lib and bin directories. I've removed the
Gemfile since this is now handled in the gemspec.

Note that this doesn't quite work when installed as a gem, since the
APP_PATH is hardcoded as a relative path in bin/kitchen. There are a
couple of other things (tmp and logs) that I need to clean up before
this is usable anyways.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>