guix: Add initial package revision
This builds and installs the package as a gem. The gem doesn't yet
work due to the tmp and log directories getting located in APP_PATH; I
still have to come up with a workaround for this.

This is also renamed to guix.scm so it will work with a bare 'guix
shell' command (no need to point it at the file).

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>