bin: Fix startup script under guix
Instead of invoking rails' 'server' command, invoke puma directly
using the rack file located under the APP_PATH.

We also now dynamically search for APP_PATH in $LOAD_PATH if it does
not exist relative to the bin directory. This happens when the app is
installed as a gem (e.g. under guix) and bin/kitchen is relocated
somewhere in the user's path.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>